Organized files and added starting page route
Changes made:
- Relocated all components into a components folder
- Moved App.js into the routes folder
- Edited index.js with the root route for this project
