+++ src/components/JikanComponent.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useState } from "react";
-import "./styles.css";
+import "../styles.css";
 
 export default function JikanComponent() {
   const [anime, setAnime] = useState({ data: {} });
